requirements.txt
requests==2.28.2
pydantic==1.10.7
PyYAML==6.0
